Change: Use ushort for vertex indices when drawing VBO
Renderer will fall over long before we hit 65 535 vertices in a face....

+++ source/ObjectViewer/NewRendererS.cs
@@ -132,7 +132,7 @@ namespace OpenBve
 				vertexData[i].Color = Color;
 			}
 
-			int[] indexData =
+			ushort[] indexData =
 			{
 				0, 1, 2, 3,
 				0, 4, 5, 1,
